Add files via upload
Just an update to some of the content I had. 

  - The footer was fixed. 
  - Edited the content on all pages. 
  - Added some 'social media contact buttons' to the 'contact' page. 
  - Added content (Logos, I designed.) to the portfolio page. 

Pros - Generally filled like this was a Frankenstein of a project but happy with the overall look of the website. 

Cons - I could add more; content, interactive elements and more of the a creative touch to the overall looks and feel of the site.